Updated Card and Map part
For Card, I made the draw and exchange method work (it doesnt update the player's money thought)

For map, I created a new array which contains all the countries. Each element of the array is 1 node (so 1 country). That element has a string player and integer army. It tells which player owns that region, and how many armies he has.